+++ app/helpers/application_helper.rb
@@ -24,7 +24,7 @@ module ApplicationHelper
 
   def league_background(league_name)
     case league_name
-    when 'English Premier League', 'Premier Liga' then 'english_premier_league.webp'
+    when 'English Premier League', 'Premier League' then 'english_premier_league.webp'
     when 'NBA' then 'nba_league.webp'
     end
   end